pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brynhild"
version = "0.1.0"
description = "A modular AI coding assistant for the terminal"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Brynhild Contributors"}
]
keywords = ["brynhild", "ai", "cli", "coding-assistant", "openrouter"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Typing :: Typed",
]
dependencies = [
    "openai>=1.0.0",
    "click>=8.0",
    "rich>=13.0",
    "textual>=0.50",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "watchfiles>=0.21",
    "tree-sitter>=0.21",
    "tree-sitter-bash>=0.21",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-dotenv>=0.5",
    "respx>=0.20",
    "mypy>=1.8",
    "pyright>=1.1",
    "black>=24.0",
    "ruff>=0.3",
    "types-PyYAML>=6.0",
]

[project.scripts]
brynhild = "brynhild.cli:main"

[project.urls]
# TODO: Update URLs when project has a public presence
Homepage = "https://example.com/brynhild"
Documentation = "https://example.com/brynhild/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
brynhild = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks cross-component integration tests",
    "system: marks multi-subsystem tests",
    "e2e: marks end-to-end CLI workflow tests",
    "live: marks tests that make actual API calls (requires API keys)",
    "ollama: marks tests that require Ollama server (set BRYNHILD_OLLAMA_HOST or OLLAMA_HOST)",
    "profiles: marks tests for model profile system",
    "ollama_local: marks tests that run against local/private Ollama server",
    "smoke: marks smoke tests that run bin/brynhild as subprocess",
]
# Load .env file for tests (pytest-dotenv)
# .env is gitignored and contains local configuration/secrets
env_files = [".env"]

[tool.coverage.run]
source = ["src/brynhild"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
show_missing = true

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
packages = ["brynhild"]

[tool.pyright]
pythonVersion = "3.11"
venvPath = "."
venv = "local.venv"
typeCheckingMode = "basic"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator suggestion - prefer explicit if/else for readability
]

[tool.ruff.lint.isort]
known-first-party = ["brynhild"]

[tool.black]
line-length = 100
target-version = ["py311"]

